Require model selection to compute total count
Consider the scenario when a PageInfo is returned indicating that a
next page exists, holding a single element. Before the client requests
that page, the element is deleted. Previously, a total count of 0 would
have been computed since the new page is now empty.
This allows to simplify Cursor.has_next_previous_page() a little.