Improve performance of boxes-for-base query (#292)
* Query location name in location field already

- avoid superfluous calls to resolve identical location name for boxes
  of same location

* Git-ignore generated GraphQL API docs

* Remove duplicated entries from .gitignore files

* Avoid redundant costly select query for pagination

If a `selection` is defined (i.e. passed into load_into_page() from a
resolver, e.g. the peewee.ModelSelect object `Box.select()` in the
Location.boxes resolver), the line
    selection = selection or model.select()
causes peewee to execute the select query (here: SELECT * FROM stock)
which is expensive and not required.
Local testing shows that the execution time for a GraphQL query like
    query { base(id: 1) { locations { boxes { elements { labelIdentifier } } } } }
can be drastically reduced