Actually store date of signature when creating beneficiary
Work-around the following issue:
The dateOfSignature field is optional in BeneficiaryCreationInput. If
not specified it is passed as None into create_beneficiary and
Beneficiary.create() but the underlying MySQL field definition dictates
NOT NULL and defines a default itself (the zero-date-string 0000-00-00
which cannot be expressed as a valid Python date object).
Hence use the value of dateOfSignature only if not None, otherwise
fallback to the MySQL default.
The situation can hopefully be resolved when the date_of_signature
field is migrated to allow NULL values.