Return only beneficiaries from the user's auth'zed bases on Tag.taggedResources
This is currently safe because the routes to get to this resolver
already enforce base-specific permissions:
- tag(id) query
- tags query
- box(labelIdentifier).tags
- beneficiary(id).tags
- beneficiaries.tags

But in the future a new GraphQL type Foo could be introduced, containing
a Tag field without enforcing base-specific permissions in the resolver.
Then by Foo.tag.taggedResources it would be possible to obtain all
beneficiaries.