Correctly configure CORS in back-end
When running the back-end tests a warning "unknown option passed to
flask-cors: origin" could be seen. It turns out that the option
correctly has to be named 'origins'. But even then the server would
accept any origins for making requests.
More research revealed that the flask-cors offers several ways of
configuring the extension: using the `CORS` wrapper, and using the
`cross_origin` decorator. Since we used both, the former took
precedence, and it's default is to allow any origins, methods, and
headers (questionable defaults, however that's exactly what the docs
state).
As a solution, the `CORS` wrapper is removed, and the `cross_origin`
decorator is applied to the one route that matters, with the correct
options passed in.