Sort history entries by ID instead of change date
The issue was that we use ordering by change datetime to sort log
entries, and then select the first one as the newest one. However the
datetime has second precision, hence when two changes occur within one
second, it's not necessarily the newest one that comes first in the
ordering. The newest entry always has the largest ID, so sorting by ID
is used now.