Make error detection for re-used QR code less strict
Since the last prod deploy, the MySQL error message for when a used QR
code is added to a box is along
    Duplicate entry '153659' for key 'qr_id_unique'
but the error detection was looking for an exact match
    Duplicate entry '153659' for key 'stock.qr_id_unique'
and, not finding a match, would leave this error unhandled.
A couple of occurrences have been reported by Sentry recently.

The error detection is now made less strict since we only have to
distinguish between an assignment of a duplicated QR code, a duplicated
box label identifier, or another unknown error.

This fix is important because the detection of a used QR code would
trigger an unknown error without telling the user what went wrong
(hence also the uncaught exceptions in Sentry). Now they're correctly
notified again about the QR code already being in use.