Perform date conversion earlier when creating agreement
From the GraphQL interface, a Date type is passed in. For the data
level, but also for the business logic, it needs to be converted into a
timezone-aware UTC datetime. This should happen as early as possible in
create_transfer_agreement(); and for clarity it's moved into a separate
method. Set default for valid_from to current time there. This
implicitly fixes a bug that occurred when validUntil was specified but
validFrom was not.