Use up-to-date data when trying to complete shipment
Updating a shipment when receiving means to update reconciled boxes as
InStock and lost boxes as NotDelivered. Finally it is checked whether
the shipment can be marked as completed.
Previously these three steps (updating reconciled boxes, updating lost
boxes, trying to complete shipment) happened in one database
transaction. For the final step, all boxes corresponding to the shipment
are selected and their states are checked. However this selection will
not reflect the most recent changes in box state because the
transaction has not yet been committed.
Now a separate transaction is used to have boxes updated first, and
then check whether the box states indicate that the shipment has been
completed.