Remove confusing index of base test data fixture
All fixtures for multiple test data instances (organisations,
shipments, boxes, ...) return a list of dicts (usually the elements
with ID 1 is at index 0 of the list, ID 2 at index 1 etc.). For bases
test data, the fixture would return a dict (key 1 for element with ID 1
etc.). This is confusing and now aligned.
Also introduce a new fixture for clarity.