Use separate database for cron tests
When the reseed-db cron test has run, the mysql_testing_database
fixture is essentially replaced by the seed. When setting up this
fixture for the next test, all tables corresponding to v2 data models
are dropped first, then re-created.
The seed however contains table definitions that are not present in the
v2 model definitions, e.g. `itemsout` which holds a FK to the sizes
table. This results in errors like
    Cannot drop table 'sizes' referenced by a foreign key constraint
    'itemsout_ibfk_8' on table 'itemsout'.