Use defaultTableConfig when refetching boxesview data after creating box
Omitting defaultTableConfig leads to the error
    Uncaught (in promise) TypeError: Cannot read properties of undefined (reading 'columnFilters')
        at Object.getColumnFilters (hooks.ts:133:34)
        at onSubmitBoxCreateForm (BoxCreateView.tsx:185:81)
        at onSubmit (BoxCreate.tsx? [sm]:133:65)
        at createFormControl.ts:1247:17
when trying to create a box without having been to BoxesView before

Reproduce by
1. Start the app
2. Directly go to /boxes/create (simulate user scanning a code directly
   after opening the app)
3. Fill the form and submit
4. Observe the error above

Why?
On submitting the create-box data, the FE wants to run a refetch-query
to get boxes-view data. For this it needs tableConfig.getColumnFilters()
but the underlying tableConfigsState is not initialized with default
values because BoxesView has not been opened before