Fix test data generation
In the routine for test data generation, tables must be created and
populated in a certain order such that independent tables are created
first. This is achieved through a hard-coded list of modules (_NAMES),
and a remainder of modules that only depend on other modules but are no
dependency of another module.
This remainder is obtained via file search (os.listdir). It contained,
among others, the packing_list_entry and the distribution_event
modules. Containing the latter is an error because it is a dependency
of the packing_list_entry.
Until recently test runs in all local dev envs and in CI have always
resulted in an ordering of the files such that the distribution_event
was set up first BY CHANCE BUT on William's environment it didn't,
resulting in the MySQL error
    Cannot add or update a child row: a foreign key constraint fails (`testingreadonly`.`distro_events_packing_list_entries`, CONSTRAINT `distro_events_packing_list_entries_ibfk_3` FOREIGN KEY (`distro_event_id`) REFERENCES `distro_events` (`id`) ON UPDATE CASCADE
upon trying to insert data.

The fix is to explicitly list the modules that are dependencies of
others in the _NAMES list.
Also the remainder of modules is sorted alphabetically such that it
behaves the same in all envs regardless of the os.listdir result.